requirements.txt
pyaudio==0.2.11
SpeechRecognition==3.10.0
googletrans==4.0.0rc1
pystray==0.19.4
Pillow==10.0.0
numpy==1.24.3
openai-whisper==20231117
torch
torchaudio
openai
